+++ languagetool-core/src/main/java/org/languagetool/rules/Rule.java
@@ -64,6 +64,11 @@ public abstract class Rule {
    */
   public Rule(ResourceBundle messages) {
     this.messages = messages;
+    if (messages != null) {
+      setCategory(Categories.MISC.getCategory(messages));  // the default, sub classes may overwrite this
+    } else {
+      setCategory(new Category(CategoryIds.MISC, "Misc"));
+    }
   }
 
   /**