+++ languagetool-core/src/test/java/org/languagetool/tools/RuleMatchAsXmlSerializerTest.java
@@ -98,7 +98,7 @@ public class RuleMatchAsXmlSerializerTest {
     assertTrue(xml.contains(">\n" +
             "<error fromy=\"44\" fromx=\"98\" toy=\"45\" tox=\"99\" ruleId=\"FAKE_ID\" msg=\"myMessage\" " +
             "replacements=\"\" context=\"...s is an test...\" contextoffset=\"8\" offset=\"8\" errorlength=\"2\" " +
-            "locqualityissuetype=\"misspelling\"/>\n" +
+            "category=\"Misc\" categoryid=\"MISC\" locqualityissuetype=\"misspelling\"/>\n" +
             "</matches>\n"));
   }
 
@@ -162,7 +162,7 @@ public class RuleMatchAsXmlSerializerTest {
     assertTrue(xml.contains(">\n" +
             "<error fromy=\"44\" fromx=\"98\" toy=\"45\" tox=\"99\" ruleId=\"FAKE_ID\" msg=\"myMessage\" " +
             "replacements=\"\" context=\"...s is an test...\" contextoffset=\"8\" offset=\"8\" errorlength=\"2\" url=\"http://server.org?id=1&amp;foo=bar\" " +
-            "locqualityissuetype=\"misspelling\"/>\n" +
+            "category=\"Misc\" categoryid=\"MISC\" locqualityissuetype=\"misspelling\"/>\n" +
             "</matches>\n"));
   }
 
@@ -180,7 +180,7 @@ public class RuleMatchAsXmlSerializerTest {
     assertTrue(xml.contains(">\n" +
             "<error fromy=\"44\" fromx=\"98\" toy=\"45\" tox=\"99\" ruleId=\"FAKE_ID\" msg=\"myMessage\" " +
             "replacements=\"\" context=\"... is &quot;an test...\" contextoffset=\"8\" offset=\"9\" errorlength=\"2\" " +
-            "locqualityissuetype=\"misspelling\"/>\n" +
+            "category=\"Misc\" categoryid=\"MISC\" locqualityissuetype=\"misspelling\"/>\n" +
             "</matches>\n"));
   }